+++ crates/starknet_api/src/test_utils/invoke.rs
@@ -2,6 +2,7 @@ use crate::calldata;
 use crate::core::{ContractAddress, Nonce};
 use crate::data_availability::DataAvailabilityMode;
 use crate::executable_transaction::InvokeTransaction as ExecutableInvokeTransaction;
+use crate::rpc_transaction::{RpcInvokeTransaction, RpcInvokeTransactionV3, RpcTransaction};
 use crate::transaction::{
     AccountDeploymentData,
     Calldata,
@@ -113,4 +114,27 @@ pub fn executable_invoke_tx(invoke_args: InvokeTxArgs) -> ExecutableInvokeTransa
     let tx = invoke_tx(invoke_args);
 
     ExecutableInvokeTransaction { tx, tx_hash }
+}
+
+pub fn rpc_invoke_tx(invoke_args: InvokeTxArgs) -> RpcTransaction {
+    if invoke_args.version != TransactionVersion::THREE {
+        panic!("Unsupported transaction version: {:?}.", invoke_args.version);
+    }
+
+    let ValidResourceBounds::AllResources(resource_bounds) = invoke_args.resource_bounds else {
+        panic!("Unspported resource bounds type: {:?}.", invoke_args.resource_bounds)
+    };
+
+    RpcTransaction::Invoke(RpcInvokeTransaction::V3(RpcInvokeTransactionV3 {
+        resource_bounds,
+        tip: invoke_args.tip,
+        calldata: invoke_args.calldata,
+        sender_address: invoke_args.sender_address,
+        nonce: invoke_args.nonce,
+        signature: invoke_args.signature,
+        nonce_data_availability_mode: invoke_args.nonce_data_availability_mode,
+        fee_data_availability_mode: invoke_args.fee_data_availability_mode,
+        paymaster_data: invoke_args.paymaster_data,
+        account_deployment_data: invoke_args.account_deployment_data,
+    }))
 }
